Add confirmation popup to VM detail
